enh: separate macros sources for 2.10 and 2.11+
